Support dynamic extrinsics parameter updates (#34)
Add support to ROS driver to modify extrinsics. The reconfigure callback updates the 
extrinsics on-board the multisense — this is important to tools such as ground surface 
modelling which run on the camera and rely on these extrinsics being set correctly. 
The callback also updates a transform member in the camera class, which allows the 
user to dynamically redraw the pointcloud position in the RVIZ window.